refactor: Refactor saveRegister function to include program information
The code changes in the `service.ts` file refactor the `saveRegister` function to include program information in the registration data. The program code is extracted from the reference string and used to retrieve the corresponding program object using the `getProgramByCode` function. If the program is not found, an error is thrown.

This refactoring improves the data integrity of the registration process by associating each registration with a specific program. It also ensures that the program information is stored correctly in the Firestore database.

Note: The commit message includes specific details about the refactoring of the `saveRegister` function and the addition of program information.